Add gray-scale display. (#8)
* Add gray-scale display.

This display type can be useful for the 16-shades
gray-scale e-paper displays.  It uses one byte per
pixel of gray-scale data.  All the primitives are
used unchanged from the true-color 24 bit display.
Almost all the code is a straight-forward copy of
the true-color code, reduced from three channels
to one.

* Doc comment improvements