Refactor text redraw box code.
When a text is changed on a TextTexture we try
to detect whether most of the string is unchanged
and only redraw the changed part.

This code is now moved to a static method so it
can be used from the V2 graphical elements.

At the same time it has been improved so it can
work for center-aligned strings where only part
of it changes.