Start adjustments for Version 2 (#47)
* Refactor to support version2.

In version 2 of the pixel-display, the textures are going
away, and will be replaced by elements.

The display as a whole will have a 2d transform (rotation),
and the individual elements (windows) will have only an internal
offset, not a full transform.

The Canvases know their offset relative to the scene, and
will eventually gain drawing operations.  For now we just
rename them, to put more emphasis on the abstract class
at the top of the hierarchy, which will be more important
going forward than the various bit depth variants.

Start to move bar code code out of the core files.

Start common.toit which will have code common to
textures and elements in the transition period.

All drivers must know their rounding requirements going
forward.

Some changes for recent SDKs.  We now require alpha 121.

* Update gold files for new SDK's better alpha blending

* feedback